Refactor the way stopping speech is handled
Canceling the wave thread seems to cause various problems, particularly if
flite is using pulse, but the ALSA backend also deadlocks or crashes
occasionally. Instead of using pthread_cancel, use a variable to indicate
that the thread should be canceled. This way, we can abort the thread only
at points where it is known to be safe. We now write audio data in small
chunks. Also, the wave thread is now responsible for closing the audio
device.
This has a negative effect in terms of latency with the ALSA backend, since
flushing takes time and we do not check for cancelation during the flush,
but hopefully this change will improve stability, and it appears to make the
pulse backend functional.